% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchPPIScore.R
\name{BenchPPIScore}
\alias{BenchPPIScore}
\title{Benchmark PPI Scores using ROC}
\usage{
BenchPPIScore(ppi, score, refInt)
}
\arguments{
\item{ppi}{Character vector of PPI strings.}

\item{score}{Numeric vector of scores associated with `ppi`.}

\item{refInt}{A list with 'TP' and 'TN' data frames for benchmarking.}
}
\value{
A `pROC::roc` object.
}
\description{
This function evaluates the performance of PPI scores against gold standard
True Positive (TP) and True Negative (TN) interactions using ROC analysis.
}
