% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_statistics.R
\name{calculate_complex_statistics}
\alias{calculate_complex_statistics}
\title{Calculate Statistics for a List of Protein Complexes}
\usage{
calculate_complex_statistics(complex_list)
}
\arguments{
\item{complex_list}{A list where each element is a character vector
representing a protein complex (i.e., a list of protein identifiers).}
}
\value{
A named numeric vector with the following statistics:
  \item{num_clusters}{Total number of complexes in the list.}
  \item{num_subunits}{Total number of unique protein subunits across all
    complexes.}
  \item{pct_dimers}{Percentage of complexes that are dimers (size 2).}
  \item{pct_trimers}{Percentage of complexes that are trimers (size 3).}
  \item{pct_large_gt10}{Percentage of complexes with more than 10 subunits.}
  \item{avg_complex_size}{Average number of subunits per complex.}
}
\description{
Computes various descriptive statistics for a given list of protein
complexes, such as the number of complexes, number of unique subunits,
average complex size, and percentages of specific N-mers.
}
\examples{
complexes <- list(
  cpx1 = c("A", "B"),
  cpx2 = c("B", "C", "D"),
  cpx3 = c("E", "F"),
  cpx4 = LETTERS[1:12]
)
stats <- calculate_complex_statistics(complexes)
print(stats)
}
