% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_intersection_matrix}
\alias{calculate_intersection_matrix}
\title{Calculate Pairwise Intersections Between Complexes}
\usage{
calculate_intersection_matrix(complexes1, complexes2, num_cores = 1)
}
\arguments{
\item{complexes1}{First list of protein complexes (character vectors)}

\item{complexes2}{Second list of protein complexes (character vectors)}

\item{num_cores}{Number of cores for parallel processing (default = 1)}
}
\value{
Numeric matrix where entry [i,j] contains the size of intersection between
        complexes1[[i]] and complexes2[[j]]
}
\description{
Computes a matrix of intersection sizes between all pairs of complexes from two lists.
Used internally for complex similarity calculations.
}
\details{
This function performs all-vs-all comparison of protein complexes using parallel processing.
For each pair of complexes (cpx1, cpx2), it calculates the size of their intersection.
The result is a matrix where entry [i,j] = length(intersect(cpx1[[i]], cpx2[[j]])).
}
\examples{
# Compare two sets of complexes
cpx1 <- list(c("A", "B", "C"), c("D", "E"))
cpx2 <- list(c("A", "B"), c("C", "D", "F"))
calculate_intersection_matrix(cpx1, cpx2)
}
\seealso{
\code{\link{calculate_jaccard_matrix}} for similarity metrics,
         \code{\link{calculate_pr_matrix}} for precision-recall scores
}
\keyword{internal}
