% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_n_cores}
\alias{.get_n_cores}
\title{Get Number of Cores for Parallelism}
\usage{
.get_n_cores()
}
\value{
Integer number of cores.
}
\description{
Internal utility to respect CRAN/Bioconductor and user settings.
}
\keyword{internal}
