% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_ppi_string}
\alias{.get_ppi_string}
\title{Get PPI String}
\usage{
.get_ppi_string(A, B)
}
\arguments{
\item{A}{Character vector of first interactors.}

\item{B}{Character vector of second interactors.}
}
\value{
Character vector of "A~B" strings.
}
\description{
Internal utility to create a normalized PPI string.
}
\keyword{internal}
