% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.normalize_ppi}
\alias{.normalize_ppi}
\title{Normalize PPI Pairs}
\usage{
.normalize_ppi(A, B)
}
\arguments{
\item{A}{Character vector of first interactors.}

\item{B}{Character vector of second interactors.}
}
\value{
A list with normalized `A` and `B`.
}
\description{
Internal utility to ensure InteractorA is alphabetically smaller than InteractorB.
}
\keyword{internal}
