% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{dummy_train_complexes}
\alias{dummy_train_complexes}
\title{Dummy Reference Complexes for Training}
\format{
A list with two elements:
\describe{
  \item{TP}{Data frame of true positive interactions.}
  \item{TN}{Data frame of true negative interactions.}
}
}
\source{
Synthetic data generated for package examples.
}
\usage{
dummy_train_complexes
}
\description{
A list containing True Positive (TP) and True Negative (TN) PPI pairs for training.
}
\keyword{datasets}
