% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_mad_threshold}
\alias{filter_by_mad_threshold}
\title{Filter Proteins by MAD Threshold}
\usage{
filter_by_mad_threshold(mean_ratios_per_protein, mad_threshold = 2.5)
}
\arguments{
\item{mean_ratios_per_protein}{Named numeric vector of mean split ratios}

\item{mad_threshold}{Number of MADs from median to use as cutoff (default = 2.5)}
}
\value{
Character vector of protein names passing filter
}
\description{
Identifies proteins whose split ratios deviate from median by specified number
of MADs (median absolute deviations). Used to find proteins with asymmetric
elution patterns.
}
\details{
Filtering process:
1. Computes median and MAD of split ratios
2. If MAD is 0 (all ratios identical), returns proteins with non-median ratios
3. Otherwise, returns proteins outside median ± (threshold * MAD)
}
\examples{
# Filter proteins with asymmetric elution
ratios <- c(A=1.2, B=0.8, C=3.5, D=1.1)
filter_by_mad_threshold(ratios, mad_threshold=2)
}
\seealso{
\code{\link{calculate_split_ratios}} for computing input ratios
}
\keyword{internal}
