% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetComplexPPI.R
\name{GetComplexPPI}
\alias{GetComplexPPI}
\title{Generate Training PPI Pairs from Protein Complexes}
\usage{
GetComplexPPI(rawCpx)
}
\arguments{
\item{rawCpx}{A list of character vectors, where each vector contains protein
IDs belonging to a complex.}
}
\value{
A list with two elements:
  \item{TP}{A data frame of True Positive PPIs.}
  \item{TN}{A data frame of True Negative PPIs.}
}
\description{
This function takes a list of protein complexes and generates all possible
True Positive (TP) pairs within the complexes and True Negative (TN) pairs
(all other possible pairs among the proteins involved).
}
