% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrtPPI.R
\name{GetPrtPPI}
\alias{GetPrtPPI}
\title{Generate Pairwise PPI Strings from Protein List}
\usage{
GetPrtPPI(prt)
}
\arguments{
\item{prt}{A character vector of protein IDs.}
}
\value{
A data.table with columns `InteractorA` and `InteractorB`.
}
\description{
This function generates all possible pairwise interactions (PPI strings)
from a list of protein IDs.
}
