% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateScore.R
\name{IntegrateScore}
\alias{IntegrateScore}
\title{Integrate Scores using a Single ML Model}
\usage{
IntegrateScore(rawScore, rawResponse, fnM, seed = 100)
}
\arguments{
\item{rawScore}{A numeric matrix of elution-based scores (exclusive of PPI IDs).}

\item{rawResponse}{A character vector of "Y" and "N" labels for training.}

\item{fnM}{Character vector specifying the machine learning method.}

\item{seed}{Optional seed for reproducibility. Default is 100.}
}
\value{
A numeric vector of integrated scores.
}
\description{
This is a lower-level integration function that trains a single ML model
using specified method `fnM`.
}
