% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_cofrac_293NE12_hcw}
\alias{load_cofrac_293NE12_hcw}
\title{Load Example CoFrac Elution Matrix (Havugimana 2012, 293NE12 HCW)}
\usage{
load_cofrac_293NE12_hcw()
}
\value{
A numeric matrix representing protein elution profiles.
}
\description{
Loads an example co-fractionation count matrix from Havugimana et al., 2012,
specifically the LTQ 293NE12 HCW dataset.
}
\examples{
if (interactive()) { # Avoid running automatically during checks if large
  cofrac_293NE12 <- load_cofrac_293NE12_hcw()
  # print(dim(cofrac_293NE12))
  # print(head(cofrac_293NE12[,1:5]))
}
}
