% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_cofrac_HeLaCE12_tcs}
\alias{load_cofrac_HeLaCE12_tcs}
\title{Load Example CoFrac Elution Matrix (Havugimana 2012, HeLaCE12 TCS)}
\usage{
load_cofrac_HeLaCE12_tcs()
}
\value{
A numeric matrix representing protein elution profiles.
}
\description{
Loads an example co-fractionation count matrix from Havugimana et al., 2012,
specifically the LTQ HeLaCE12 TCS dataset.
}
\examples{
if (interactive()) {
  cofrac_HeLaCE12 <- load_cofrac_HeLaCE12_tcs()
  # print(dim(cofrac_HeLaCE12))
}
}
