% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_pred_cpx_havugimana}
\alias{load_pred_cpx_havugimana}
\title{Load Example Predicted Complexes (Havugimana 2012 subset)}
\usage{
load_pred_cpx_havugimana()
}
\value{
A list where each element is a character vector of protein
  identifiers representing a complex.
}
\description{
Loads an example list of predicted protein complexes,
relevant to the Havugimana et al., 2012 study context.
}
\examples{
if (interactive()) {
  pred_complexes <- load_pred_cpx_havugimana()
  # print(length(pred_complexes))
  # if (length(pred_complexes) > 0) print(head(pred_complexes[[1]])))
}
}
