% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearning.R
\name{MachineLearning}
\alias{MachineLearning}
\title{Integrate Scores via Machine Learning}
\usage{
MachineLearning(rawScore, refIntTrain, fnMachine)
}
\arguments{
\item{rawScore}{A data frame containing PPI pairs and their elution-based scores.}

\item{refIntTrain}{A list containing 'TP' and 'TN' data frames for training.}

\item{fnMachine}{Character vector of machine learning methods (e.g., "xgbTree", "rf").}
}
\value{
A data frame with the original scores and an additional 'SMED' column for the integrated score.
}
\description{
This function integrates multiple elution-based scores using machine learning
models (passed via `fnMachine`) to produce a final integrated PPI score.
}
