% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEluteRidge.R
\name{PlotEluteRidge}
\alias{PlotEluteRidge}
\title{Plot Elution Profiles as Ridge Plot}
\usage{
PlotEluteRidge(dat)
}
\arguments{
\item{dat}{A numeric matrix or data frame of protein elution profiles.}
}
\value{
A ggplot object.
}
\description{
This function visualizes protein elution profiles using a ridge plot, where
each protein's intensity across fractions is shown as a density-like ridge.
}
