% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotROCobjs.R
\name{PlotROCobjs}
\alias{PlotROCobjs}
\title{Plot Multiple ROC Curves}
\usage{
PlotROCobjs(listObjs, sn, colorplatte = RColorBrewer::brewer.pal(8, "Dark2"))
}
\arguments{
\item{listObjs}{A list of ROC objects (from `pROC::roc`).}

\item{sn}{Character title for the plot.}

\item{colorplatte}{Character vector of colors for the curves.}
}
\value{
A ggplot object.
}
\description{
This function plots multiple ROC curves on a single plot for comparison using ggplot2.
}
