% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUpset.R
\name{PlotUpset}
\alias{PlotUpset}
\title{Plot PPI Overlap using UpSet Plot}
\usage{
PlotUpset(listTerms, n_interactions = 20)
}
\arguments{
\item{listTerms}{A named list where each element is a character vector of
PPI strings (e.g., "ProtA~ProtB").}

\item{n_interactions}{Maximum number of intersections to show. Default is 20.}
}
\value{
A ggplot object.
}
\description{
This function visualizes the overlap between different sets of protein-protein
interactions using an UpSet plot.
}
