% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_redundancy.R
\name{remove_complex_redundancy}
\alias{remove_complex_redundancy}
\title{Remove Complex Redundancy via Hierarchical Clustering}
\usage{
remove_complex_redundancy(
  raw_complex_list,
  hclust_method = "ward.D2",
  cut_height = 0.2
)
}
\arguments{
\item{raw_complex_list}{A list where each element is a character vector
of protein identifiers representing a complex.}

\item{hclust_method}{The agglomeration method to be used for hierarchical
clustering. See `hclust` (e.g., "ward.D2", "complete"). Default "ward.D2".}

\item{cut_height}{Numeric, the height at which to cut the dendrogram to
form clusters of complexes. Complexes within the same cluster will be
merged. Default 0.2. This threshold applies to the heights derived from
Euclidean distances of Jaccard similarity profiles.}
}
\value{
A list of refined complexes. Each element is a character vector of
  unique protein identifiers, and names are generated based on cluster IDs
  (e.g., "CID_1").
}
\description{
Reduces redundancy in a list of protein complexes. It calculates pairwise
Jaccard indices between all complexes, performs hierarchical clustering on
Euclidean distances of these Jaccard similarity profiles,
and then merges complexes that fall into the same cluster by uniting their
protein members.
}
\examples{
cplx_list <- list(
  c1 = c("A", "B", "C"),
  c2 = c("B", "C", "D"), # Similar to c1
  c3 = c("C", "D", "E"), # Similar to c2
  c4 = c("F", "G"),      # Different
  c5 = c("A", "B", "C", "D") # Superset of c1 and c2
)
if (requireNamespace("proxy", quietly = TRUE)) {
  refined_cplx <- remove_complex_redundancy(cplx_list, cut_height = 0.5)
  # print(refined_cplx)
}
}
