% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveComplexRedundance.R
\name{RemoveComplexRedundance}
\alias{RemoveComplexRedundance}
\title{Remove Redundant Protein Complexes}
\usage{
RemoveComplexRedundance(rawCpx)
}
\arguments{
\item{rawCpx}{A list of character vectors, each representing a protein complex.}
}
\value{
A list of unique protein complexes with renamed IDs.
}
\description{
This function merges redundant protein complexes in a list based on their
Jaccard similarity and hierarchical clustering.
}
