% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_cofrac_HeLaNE45_tcs}
\alias{load_cofrac_HeLaNE45_tcs}
\title{Load Example CoFrac Elution Matrix (Havugimana 2012, HeLaNE45 TCS)}
\usage{
load_cofrac_HeLaNE45_tcs()
}
\value{
A numeric matrix representing protein elution profiles.
}
\description{
Loads an example co-fractionation count matrix from Havugimana et al., 2012,
specifically the LTQ HeLaNE45 TCS dataset.
}
\examples{
if (interactive()) {
  cofrac_HeLaNE45 <- load_cofrac_HeLaNE45_tcs()
  # print(dim(cofrac_HeLaNE45))
}
}
