% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{load_reference_corum_havugimana}
\alias{load_reference_corum_havugimana}
\title{Load Example CORUM Reference Complexes (Havugimana 2012 subset)}
\usage{
load_reference_corum_havugimana()
}
\value{
A list where each element is a character vector of protein
  identifiers representing a complex.
}
\description{
Loads an example list of CORUM reference protein complexes,
relevant to the Havugimana et al., 2012 study context.
}
\examples{
if (interactive()) {
  corum_complexes <- load_reference_corum_havugimana()
  # print(length(corum_complexes))
  # if (length(corum_complexes) > 0) print(head(corum_complexes[[1]])))
}
}
