% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotROCobjs.R
\name{PlotROCobjs}
\alias{PlotROCobjs}
\title{Plot Multiple ROC Curves}
\usage{
PlotROCobjs(
  listObjs,
  sn,
  colorplatte = RColorBrewer::brewer.pal(8, "Dark2"),
  txtX = 0.55,
  txtY = 0.6
)
}
\arguments{
\item{listObjs}{A list of ROC objects (from `pROC::roc`).}

\item{sn}{Character title for the plot.}

\item{colorplatte}{Character vector of colors for the curves.}

\item{txtX}{x-coordinate for the legend text.}

\item{txtY}{y-coordinate for the legend text.}
}
\value{
None (invokes a plot).
}
\description{
This function plots multiple ROC curves on a single plot for comparison.
}
