% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetainElutionProtein.R
\name{RetainElutionProtein}
\alias{RetainElutionProtein}
\title{Identify High-Confidence Eluting Proteins}
\usage{
RetainElutionProtein(m, nmad, n_boot = 1000)
}
\arguments{
\item{m}{A numeric matrix of elution profiles.}

\item{nmad}{Number of Median Absolute Deviations (MAD) from the median to
set the threshold.}

\item{n_boot}{Number of bootstrap repetitions. Default is 1000.}
}
\value{
A character vector of protein IDs that pass the filter.
}
\description{
This function identifies proteins that show non-random elution patterns
using a bootstrapping approach to compare protein intensities across
random fraction splits.
}
